[qtmozembed] Move chrome/chromeless mode logic to C++
Application developer declarates chromeGestrureThreshold
and chromeGestureEnabled properties. When chromeGestureThreshold
is exceeded chrome property is updated. When content is moved
from bottom to top and threshold exceeds chrome property is disabled.
Chrome property is enabled when panning/flicking
from top to bottom and chromeGestureThreshold is exceeded.

This commit also adds moving property that tells whether content
is moving or not.