[qtmozembed] Update view size always when context info is updated
Previously there was a chance that QGraphicsMozViewPrivate::UpdateViewSize
did not update EmbedLiteView::SetGLViewPortSize because mHasContext was still
false. This in turn led to a crash when EmbedLiteCompositorParent::RenderGL()
tried to publish a frame that was an empty.

This commit adds a updateViewSize signal that is connected to the
QGraphicsMozViewPrivate::UpdateViewSize (normal queued connection). By doing
this QuickMozView::createThreadRenderObject can update mHasContext
and EmbedLiteView::SetGLViewPortSize.