[qtmozembed] Add initialize method to complete instantiation. Contributes JB#27557
The embedlite view is created when initialize() is called. This
needs to be called as a last step when instantiating a page.

When the page is wrapped inside a qml component we cannot
pass additional argument to the QQmlComponent::beginCreate.
Once state information is set for the instance, before calling
QQmlComponent::completeCreate complete initialization by
calling initialize().